requirements.txt
Flask
python-dotenv
gitpython
chromadb
sentence-transformers
langchain
google-generativeai
python-jenkins
requests